pyproject.toml
[tool.poetry]
name = "slack_cleanup"
version = "0.1.0"
description = ""
authors = ["Nikolas Papaioannou <nikolas@mnemonic.no>"]

[tool.poetry.dependencies]
python = "^3.9"
slack-sdk = "^3.15.1"
PyYAML = "^6.0"
typer = "^0.4.0"
pydantic = "^1.9.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
pytest-order = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
